api-gateway/requirements.txt
flask==2.0.3
flask-httpauth==4.7.0
redis==4.1.0
requests==2.26.0
pyjwt==2.3.0
python-dotenv==0.19.0
gunicorn==20.1.0
werkzeug==2.0.3
